+++ source/ntfsfile.c
@@ -219,14 +219,14 @@ int ntfs_open_r (struct _reent *r, void *fileStruct, const char *path, int flags
     // Unlock
     ntfsUnlock(file->vd);
 
-    return (int)(s64) fileStruct;
+    return (int)(s64)(intptr_t) fileStruct;
 }
 
 int ntfs_close_r (struct _reent *r, int fd)
 {
     ntfs_log_trace("fd %p\n", (void *) fd);
 
-    ntfs_file_state* file = STATE(((s64) fd));
+    ntfs_file_state* file = STATE(((intptr_t)(s64) fd));
 
     // Sanity check
     if (!file || !file->vd) {
@@ -259,7 +259,7 @@ ssize_t ntfs_write_r (struct _reent *r, int fd, const char *ptr, size_t len)
 {
     ntfs_log_trace("fd %p, ptr %p, len %u\n", (void *) fd, ptr, len);
 
-    ntfs_file_state* file = STATE(((s64) fd));
+    ntfs_file_state* file = STATE(((intptr_t)(s64) fd));
     ssize_t written = 0;
     off_t old_pos = 0;
 
@@ -325,7 +325,7 @@ ssize_t ntfs_read_r (struct _reent *r, int fd, char *ptr, size_t len)
 {
     ntfs_log_trace("fd %p, ptr %p, len %u\n", (void *) fd, ptr, len);
 
-    ntfs_file_state* file = STATE(((s64) fd));
+    ntfs_file_state* file = STATE(((intptr_t)(s64) fd));
     ssize_t read = 0;
 
     // Sanity check
@@ -384,7 +384,7 @@ off_t ntfs_seek_r (struct _reent *r, int fd, off_t pos, int dir)
 {
     ntfs_log_trace("fd %p, pos %llu, dir %i\n", (void *) fd, pos, dir);
 
-    ntfs_file_state* file = STATE(((s64) fd));
+    ntfs_file_state* file = STATE(((intptr_t)(s64) fd));
     off_t position = 0;
 
     // Sanity check
@@ -413,7 +413,7 @@ s64 ntfs_seek64_r (struct _reent *r, int fd, s64 pos, int dir)
 {
     ntfs_log_trace("fd %p, pos %llu, dir %i\n", (void *) fd, pos, dir);
 
-    ntfs_file_state* file = STATE(((s64) fd));
+    ntfs_file_state* file = STATE(((intptr_t)(s64) fd));
     s64 position = 0;
 
     // Sanity check
@@ -442,7 +442,7 @@ int ntfs_fstat_r (struct _reent *r, int fd, struct stat *st)
 {
     ntfs_log_trace("fd %p\n", (void *) fd);
 
-    ntfs_file_state* file = STATE(((s64) fd));
+    ntfs_file_state* file = STATE(((intptr_t)(s64) fd));
     int ret = 0;
 
     // Sanity check
@@ -467,7 +467,7 @@ int ntfs_ftruncate_r (struct _reent *r, int fd, off_t len)
 {
     ntfs_log_trace("fd %p, len %llu\n", (void *) fd, (u64) len);
 
-    ntfs_file_state* file = STATE(((s64) fd));
+    ntfs_file_state* file = STATE(((intptr_t)(s64) fd));
 
     // Sanity check
     if (!file || !file->vd || !file->ni || !file->data_na) {
@@ -534,7 +534,7 @@ int ntfs_fsync_r (struct _reent *r, int fd)
 {
     ntfs_log_trace("fd %p\n", (void *) fd);
 
-    ntfs_file_state* file = STATE(((s64) fd));
+    ntfs_file_state* file = STATE(((intptr_t)(s64) fd));
     int ret = 0;
 
     // Sanity check
